FinalHW/requirements.txt
apache-airflow==2.5.0
apache-airflow-providers-postgres==5.3.0
apache-airflow-providers-mysql==3.3.0
psycopg2-binary==2.9.5
mysql-connector-python
Faker==18.3.1
pandas